Create week2.py
This code extends the neural network setup from the previous example (Workshop 1 - Part 1) by adding complexity to the model architecture and including a regularization technique called Dropout. This model is also designed to classify handwritten digits from the MNIST dataset. The steps and the model's architecture are more advanced compared to the first part, adds non-linearality.